.table td,
  .table th,
  .table thead th {
    vertical-align: middle !important;
  }

  .table thead th {
    height: 50px !important;
  }

  .select2-container {
  width: 100% !important;
  }
</style>

<div class="card rect-border">
  <div class="card-body">
    <table class="table table-striped mt-4 mb-4">
      <thead>
        <tr>
          <th scope="col">No.</th>
          <th scope="col">Image</th>
          <th scope="col">Name</th>
          <th scope="col">Skills</th>
          <th scope="col">Gaji</th>
          <th scope="col">Expetasi Gaji</th>
          <th scope="col">Action</th>
        </tr>
      </thead>
      <tbody id="container">
        @foreach ($data as $talent)
        <tr>
          <td>{{($data->currentPage()-1) * $data->perPage() + $loop->iteration}}</td>
          <td>
            <img src="{{url('/img/avatar/noimage.jpg')}}" style="width: 50px; height:50px;" alt="">
          </td>
          <?php $result = substr($talent->name, 0, 1) . preg_replace('/[^@]/', '*', substr($talent->name, 1));?>
          <td style="max-width: 10rem;">{{$result}}</td>
          <td style="max-width: 400px">
            @foreach ( $talent->talent_skill()->get() as $row )
            <?php

              if ( $row->st_skill_verified == "YES")
              {
                $badge = 'success';
              }
              else
              {
                $badge = 'light';
              }
              $skill = $row->skill()->first();
            ?>
            <span class="badge badge-{{$badge}}">
              {{$skill->skill_name}}
            </span>
            @endforeach
          </td>
          <td>
            @if (!empty($talent->gaji))
            {{ $talent->gaji }}
            @else
            -
            @endif
          </td>
          <td>
            @if (!empty($talent->expetasi))
            {{ $talent->expetasi }}
            @else
            -
            @endif
          </td>
          <td style="min-width:200px">
            @if (!empty($talent->user_id))
            <a href="{{url('/profile/'.encrypt_custom($talent->user_id))}}" class="btn btn-info" target="_blank">
              <i class="fa fa-info"></i>
            </a>
            @else
            <button type="button" data-toggle="modal" data-target="#detail-user" class="btn btn-info" target="_blank">
              <i class="fa fa-info"></i>
            </button>
            @endif
            <button class="btn btn-sm btn-primary rounded" data-target="#modal-offer" data-id="{{$talent->talent_id}}"
              data-toggle="modal">Make A Request</button>
          </td>
        </tr>
        @endforeach
      </tbody>
    </table>
  </div>
</div>

<div class="modal fade" id="detail-user">
  <div class="modal-dialog">
    <div class="modal-content">
      <div class="modal-header">
        <h5 class="modal-title font-weight-bold">Caution Information</h5>
        <button type="button" class="close" data-dismiss="modal" aria-label="Close">
          <span aria-hidden="true">&times;</span>
        </button>
      </div>
      <div class="modal-body">
        <p>This account does not have detailed information</p>
      </div>
    </div>
  </div>
</div>

<style>
  .pagination {
    float: right;
  }
</style>

{{$data->links()}}

<input type="button" value="go" id="go">

<div style="clear: both;"></div>

<script>
  $(document).ready(function () {
    // list.forEach(function(item,index)
    // {
    //     $(".id-"+item).prop('checked', true);
    // });

    $(".select-all").click(function () {
      $('input:checkbox').not(this).prop('checked', this.checked);
      $(".pilih").each(function () {
        id = $(this).data("id");
        pilih(id);

      });
    });
  })
</script>

{{-- Modal Offer Langsung --}}
<div class="modal fade" id="modal-offer">
  <div class="modal-dialog modal-dialog-centered">
    <div class="modal-content">
      <div class="modal-header">
        <div class="modal-title">Add To My Request</div>
        <button type="button" class="close" data-dismiss="modal">&times;</button>
      </div>
      <form method="post" enctype="multipart/form-data" class="request-offer">
      @csrf
      <div class="modal-body">
          <div class="form-group">
            <p class="text-center" id="note-need"></p>
            <div class="row justify-content-center mt-4">
              <div class="col-md-10">
                <select class="form-control" id="request" name="name_request"></select>
              </div>
            </div>
          </div>
      </div>
      <div class="modal-footer">
        <div class="nav nav-pills pull-right">
          <button type="submit" class="btn btn-success rounded">Add</button>
        </div>
      </div>
      </form>
    </div>
  </div>
</div>
{{-- End Modal Offer Langsung --}}

<script>
  $(document).ready(function () {
    $('#modal-offer').on('show.bs.modal', function (e) {
      var talent_id = e.relatedTarget.dataset.id;
      const _url = `{{ url('company/dashboard/addTalentReq?talent_id=${talent_id}') }}`;
      $('.request-offer').attr('action', _url);
    })

    $('#request').on('change', function (e) {
      var value = $(this).val();
      const _url = `{{ url('company/dashboard/getInfoReq?id_request=${value}') }}`
      $.ajax({
        url: _url,
        success: function (data) {
          var content = `<strong>${data['hired']}/${data['need']}</strong> dari total <strong>${data['total']}</strong> Talent`;
          $("#note-need").html(content);
        }
      });
    });

    loadSelect();

  })

  function loadSelect() {
    $("#request").select2({
      theme: "bootstrap",
      placeholder: "Pilih Request",
      ajax:{
        url: '{{route("json.comp_req.company")}}',
        dataType: 'json',
        delay: 250,
        processResults: function (data) {
            var results = [];
            $.each(data, function (index, item) {
                results.push({
                    id: item.id,
                    text: item.value,
                });
            });
            return {
                results: results
            }
        },
        cache: false
      },
    });
  }

</script>

+++ app/Http/Controllers/TalentNewController.php
@@ -34,27 +34,26 @@ class TalentNewController extends Controller
     public function show()
     {
         $total  = DB::table('talent')->count();
-        $member  = DB::table('talent')->where("user_id",">","0")->count();
-        $nonmember  = DB::table('talent')->where("user_id","0")->count();
-        $invitation  = DB::table('talent')->where("talent_mail_invitation",">","0")->count();
-        return view('admin.talent.home',compact('total','member','nonmember','invitation'));
+        $member  = DB::table('talent')->where("user_id", ">", "0")->count();
+        $nonmember  = DB::table('talent')->where("user_id", "0")->count();
+        $invitation  = DB::table('talent')->where("talent_mail_invitation", ">", "0")->count();
+        return view('admin.talent.home', compact('total', 'member', 'nonmember', 'invitation'));
     }
 
     function mail(Request $request)
     {
-        $talent_id = $request->id ;
-        $talent_log = Talent_log::orderBy("id","DESC");
-        $talent = null ; 
+        $talent_id = $request->id;
+        $talent_log = Talent_log::orderBy("id", "DESC");
+        $talent = null;
 
-        if ( $talent_id>0)
-        {
+        if ($talent_id > 0) {
             $talent = Talent::findOrFail($talent_id);
-            $talent_log->where('tl_talent_id',$talent_id);
+            $talent_log->where('tl_talent_id', $talent_id);
         }
 
-        $talent_log = $talent_log->paginate(5); 
+        $talent_log = $talent_log->paginate(5);
 
-        return view("admin.talent.mail",compact('talent_log','talent'));
+        return view("admin.talent.mail", compact('talent_log', 'talent'));
     }
 
     public function table(Request $request)
@@ -66,9 +65,8 @@ class TalentNewController extends Controller
 
         // dd($request->id) ; 
 
-        if ( $request->talent_id )
-        {
-            $talent_logs->where("tl_talent_id",$request->talent_id);
+        if ($request->talent_id) {
+            $talent_logs->where("tl_talent_id", $request->talent_id);
         }
 
         if ($request->mailtype) // klo type kosong , ga masuk , klo isi masuk 
@@ -148,17 +146,17 @@ class TalentNewController extends Controller
         $content = str_replace("#name#", $talent->talent_name, $request->content);
         $data['content'] = str_replace("#nama#", $talent->talent_name, $content);
 
-        $insert = new Talent_log; 
-        $id = $insert->log($request->type,$talent_id,['desc'=>'dikirim dari new list talent']); 
-        $data['id'] = $id ; 
+        $insert = new Talent_log;
+        $id = $insert->log($request->type, $talent_id, ['desc' => 'dikirim dari new list talent']);
+        $data['id'] = $id;
 
-        Mail::send($view, $data, function ($message) use ($talent,$request,$judul) {
+        Mail::send($view, $data, function ($message) use ($talent, $request, $judul) {
             $message->from('dodi@upscale.id', $request->sender);
             $message->to($talent->talent_email);
             $message->subject($judul);
         });
 
-        
+
 
         return response()->json(['status' => 1, 'email' => $talent->talent_email, 'message' => 'send berhasil']);
 
@@ -181,8 +179,8 @@ class TalentNewController extends Controller
             $content = 'ini adalah halaman content ';
             $sender = 'dodi';
         }
-        $id=0  ;
-        return view('mail.' . $view, compact('talent', 'content', 'sender','id'));
+        $id = 0;
+        return view('mail.' . $view, compact('talent', 'content', 'sender', 'id'));
     }
 
     public function paginate_data(Request $request)
@@ -191,85 +189,100 @@ class TalentNewController extends Controller
 
         // if ($request->ajax()) {
 
-            //SELECT BUILDER START
-            $default_query = "*,users.id as user_id, users.email as member_email, users.created_at as member_date, DATEDIFF(talent_start_career, now()) as pengalaman";
-            
-            if ( $request->jumlah_apply_jobs)
-            {
-                $default_query .=", COUNT(jobs_apply_id) as jumlah_apply_jobs"; 
-            }
-            else
-            {
-                $default_query .=", '-' as jumlah_apply_jobs"; 
-            }
-            $data = Talent::select(DB::raw($default_query));
-            //SELECT BUILDER END 
+        //SELECT BUILDER START
+        $default_query = "*,users.id as user_id, users.email as member_email, users.created_at as member_date, DATEDIFF(talent_start_career, now()) as pengalaman";
 
-            //JOIN BUILDER START
-            $data->join("users","talent.user_id","=","users.id","LEFT");
-            if ( $request->jumlah_apply_jobs || $request->apply == 'yes')
-            {
-                $data->join('jobs_apply','jobs_apply.jobs_apply_talent_id','=','talent.talent_id',"LEFT");
+        if ($request->jumlah_apply_jobs) {
+            $default_query .= ", COUNT(jobs_apply_id) as jumlah_apply_jobs";
+        } else {
+            $default_query .= ", '-' as jumlah_apply_jobs";
+        }
+        $data = Talent::select(DB::raw($default_query));
+        //SELECT BUILDER END 
+
+        //JOIN BUILDER START
+        $data->join("users", "talent.user_id", "=", "users.id", "LEFT");
+        if ($request->jumlah_apply_jobs || $request->apply == 'yes') {
+            $data->join('jobs_apply', 'jobs_apply.jobs_apply_talent_id', '=', 'talent.talent_id', "LEFT");
+        }
+        //JOIN BULDER END 
+
+        if ($request->talent_id) {
+            $data->where("talent_id", "LIKE", "%" . $request->talent_id . "%");
+        }
+        if ($request->talent_name) {
+            $data->where("talent_name", "LIKE", "%" . $request->talent_name . "%");
+        }
+        if ($request->talent_phone) {
+            $data->where("talent_phone", "LIKE", "%" . $request->talent_phone . "%");
+        }
+        if ($request->talent_email) {
+            $data->where("talent_email", "LIKE", "%" . $request->talent_email . "%");
+        }
+        if ($request->talent_address) {
+            $data->where("talent_address", "LIKE", "%" . $request->talent_address . "%");
+        }
+        if ($request->talent_onsite_jogja) {
+            $data->where("talent_onsite_jogja", $request->talent_onsite_jogja);
+        }
+        if ($request->talent_onsite_jakarta) {
+            $data->where("talent_onsite_jakarta", $request->talent_onsite_jakarta);
+        }
+        if ($request->talent_isa) {
+            $data->where("talent_isa", $request->talent_isa);
+        }
+        if ($request->talent_focus) {
+            $data->where("talent_focus", "LIKE", "%" . $request->talent_focus . "%");
+        }
+
+        if ($request->min_pengalaman) {
+            $data->where("pengalaman", ">", $request->pengalaman);
+        }
+        if ($request->max_pengalaman) {
+            $data->where("pengalaman", "<", $request->pengalaman);
+        }
+
+        if ($request->min_gaji_jogja) {
+            $data->where("talent_salary_jogja", ">=", $request->min_gaji_jogja);
+        }
+        if ($request->max_gaji_jogja) {
+            $data->where("talent_salary_jogja", "<=", $request->max_gaji_jogja);
+        }
+
+        if ($request->status_member == "member") {
+            $data->where("users.email", "!=", "");
+        }
+
+        if ($request->status_member == "non-member") {
+            $data->where("users.email", "=", null);
+        }
+
+        if ($request->cv == 'yes') {
+            $data->where("talent_cv_update", "!=", null);
+        }
+
+        if ($request->apply == 'yes') {
+            $data->having(DB::raw('COUNT(jobs_apply_id)'), '>', 0);
+        }
+
+        if ($request->order != '') {
+            $ar = explode(",", $request->order);
+            foreach ($ar as $row) {
+                $data->orderBy($row, "DESC");
             }
-            //JOIN BULDER END 
+        } else {
+            $data->orderBy("talent_id", "DESC");
+        }
+        $data->groupBy("talent_id");
+        $data = $data->paginate(5);
 
-            if ( $request->talent_id ) {$data->where("talent_id","LIKE","%".$request->talent_id."%"); }
-            if ( $request->talent_name ) {$data->where("talent_name","LIKE","%".$request->talent_name."%"); }
-            if ( $request->talent_phone ) {$data->where("talent_phone","LIKE","%".$request->talent_phone."%"); }
-            if ( $request->talent_email ) {$data->where("talent_email","LIKE","%".$request->talent_email."%"); }
-            if ( $request->talent_address ) {$data->where("talent_address","LIKE","%".$request->talent_address."%"); }
-            if ( $request->talent_onsite_jogja ) {$data->where("talent_onsite_jogja",$request->talent_onsite_jogja); }
-            if ( $request->talent_onsite_jakarta ) {$data->where("talent_onsite_jakarta",$request->talent_onsite_jakarta); }
-            if ( $request->talent_isa ) {$data->where("talent_isa",$request->talent_isa); }
-            if ( $request->talent_focus ) {$data->where("talent_focus","LIKE","%".$request->talent_focus."%"); }
-            
-            if ( $request->min_pengalaman ) {$data->where("pengalaman",">",$request->pengalaman); }
-            if ( $request->max_pengalaman ) {$data->where("pengalaman","<",$request->pengalaman); }
-            
-            if ( $request->min_gaji_jogja ) {$data->where("talent_salary_jogja",">=",$request->min_gaji_jogja); }
-            if ( $request->max_gaji_jogja ) {$data->where("talent_salary_jogja","<=",$request->max_gaji_jogja); }
+        //$query = DB::getQueryLog(); dd($query);
 
-            if ($request->status_member == "member") {
-                $data->where("users.email", "!=", "");
-            }
-
-            if ( $request->status_member == "non-member" )
-            {
-                $data->where("users.email","=",null);
-            }
-
-            if ( $request->cv == 'yes' )
-            {
-                $data->where("talent_cv_update","!=",null);
-            }
-
-            if ( $request->apply == 'yes' )
-            {
-                $data->having(DB::raw('COUNT(jobs_apply_id)'),'>',0);
-            }
-
-            if ( $request->order != '' )
-            {
-                $ar = explode(",",$request->order);
-                foreach ( $ar as $row)
-                {
-                    $data->orderBy($row,"DESC");
-                }
-            }
-            else
-            {
-                $data->orderBy("talent_id","DESC");
-            }
-            $data->groupBy("talent_id");
-            $data = $data->paginate(5);
-
-            //$query = DB::getQueryLog(); dd($query);
-
-            return view('admin.talent.table',compact('data'))->render();
+        return view('admin.talent.table', compact('data'))->render();
         // }
     }
 
-    public function export_excel ()
+    public function export_excel()
     {
         return Excel::download(new TalentExport, 'talent.xlsx');
     }
@@ -284,18 +297,19 @@ class TalentNewController extends Controller
     public function del(Request $request)
     {
         $delid = $request->input('delid');
-        foreach ($delid as $row) {
-            $talent = Talent::find($row);
+        if ($delid) {
+            foreach ($delid as $row) {
+                $talent = Talent::find($row);
 
-            //menghapus semua data di table user yg berelasi
-            if ( $user = User::find($talent->user_id) ) 
-            {
+                //menghapus semua data di table user yg berelasi
+                if ($user = User::find($talent->user_id)) {
 
-                $user->delete() ; 
-            } 
+                    $user->delete();
+                }
 
-            //delete 
-            Talent::where('talent_id', $row)->delete();
+                //delete 
+                Talent::where('talent_id', $row)->delete();
+            }
         }
         return back()->with('success', 'Selected Talent has been deleted successfully');
     }
@@ -308,17 +322,17 @@ class TalentNewController extends Controller
     public function insertData(Request $request)
     {
 
-    
-        
+
+
 
 
 
         $validation = $request->validate([
-            'nama'=>'required|string|max:150',
-            'email'=>'required|string|email|max:100|unique:users',
-            'password'=>'max:150|required|required_with:confirmpass|same:confirmpass',
-            'confirmpass'=>'max:150',
-            'username'=>'required|string|unique:users,username|max:150',
+            'nama' => 'required|string|max:150',
+            'email' => 'required|string|email|max:100|unique:users',
+            'password' => 'max:150|required|required_with:confirmpass|same:confirmpass',
+            'confirmpass' => 'max:150',
+            'username' => 'required|string|unique:users,username|max:150',
             // 'gender'=>'required',
             // 'alamat'=>'required',
             // 'phone'=>'required|string|max:30',
@@ -348,21 +362,21 @@ class TalentNewController extends Controller
 
         ]);
 
-        
+
 
         $user_id = DB::table('users')->insertGetId([
-                'name' => $request->nama,
-                'username' => $request->username,
-                'email' => $request->email,
-                'password' => $request->password,
-                'level' => isset($request->level)?$request->level:"undefined",
+            'name' => $request->nama,
+            'username' => $request->username,
+            'email' => $request->email,
+            'password' => $request->password,
+            'level' => isset($request->level) ? $request->level : "undefined",
         ]);
 
 
-        
+
 
         $idTalent = DB::table('talent')->insertGetId([
-            'user_id' => $user_id ,
+            'user_id' => $user_id,
             'talent_name' => $request->nama,
             'talent_email' => $request->email,
             'talent_gender' => isset($request->gender) ? $request->gender : '',
@@ -371,7 +385,7 @@ class TalentNewController extends Controller
             'talent_birth_date' => $request->birthdate,
             'talent_place_of_birth' => $request->birthplace,
             'talent_martial_status' => isset($request->martialstatus) ? $request->martialstatus : '',
-            'talent_current_address' => isset($request->currentaddress)?$request->currentaddress:'',
+            'talent_current_address' => isset($request->currentaddress) ? $request->currentaddress : '',
             'talent_condition' => $request->condition,
             'talent_salary' => $request->salary,
             'talent_focus' => $request->focus,
@@ -415,63 +429,56 @@ class TalentNewController extends Controller
 
 
         return redirect('admin/talent/list/insert')->with('success', 'Data Talent Berhasil dimasukkan.');
-       
     }
 
 
 
-    public function import(Request $request) 
-	{
-		// validasi
-		$this->validate($request, [
-			'file' => 'required|mimes:csv,xls,xlsx'
-		]);
- 
-		// menangkap file excel
-		$file = $request->file('file');
- 
-		// membuat nama file unik
-		$nama_file = rand().$file->getClientOriginalName();
- 
-		// upload ke folder file_importExcel di dalam folder public
-		$file->move('file_importExcel',$nama_file);
- 
-        
-		// import data
-        try 
-        {
-            $cek = Excel::import(new TalentImport, public_path('/file_importExcel/'.$nama_file));
-            File::delete(public_path('file_importExcel/'.$nama_file));
+    public function import(Request $request)
+    {
+        // validasi
+        $this->validate($request, [
+            'file' => 'required|mimes:csv,xls,xlsx'
+        ]);
 
-            Session::flash('sukses','Data Excel Berhasil Diimport!');
-            
-        } 
-        catch (\Maatwebsite\Excel\Validators\ValidationException $e) 
-        {
+        // menangkap file excel
+        $file = $request->file('file');
+
+        // membuat nama file unik
+        $nama_file = rand() . $file->getClientOriginalName();
+
+        // upload ke folder file_importExcel di dalam folder public
+        $file->move('file_importExcel', $nama_file);
+
+
+        // import data
+        try {
+            $cek = Excel::import(new TalentImport, public_path('/file_importExcel/' . $nama_file));
+            File::delete(public_path('file_importExcel/' . $nama_file));
+
+            Session::flash('sukses', 'Data Excel Berhasil Diimport!');
+        } catch (\Maatwebsite\Excel\Validators\ValidationException $e) {
             $failures = $e->failures();
-            foreach ($failures as $failure) 
-            {
-                 // $failure->row(); // row that went wrong
-                 // $failure->attribute(); // either heading key (if using heading row concern) or column index
-                 // $failure->errors(); // Actual error messages from Laravel validator
-                 // $failure->values(); // The values of the row that has failed.
+            foreach ($failures as $failure) {
+                // $failure->row(); // row that went wrong
+                // $failure->attribute(); // either heading key (if using heading row concern) or column index
+                // $failure->errors(); // Actual error messages from Laravel validator
+                // $failure->values(); // The values of the row that has failed.
                 // dd($failure->values()['email']) ; 
-                Session::flash('gagal',"gagal import database ".$failure->values()['email']." sudah ada di database");
+                Session::flash('gagal', "gagal import database " . $failure->values()['email'] . " sudah ada di database");
             }
-            
         }
- 
+
         // File::delete(public_path('data_file/'.$nama_file->file));     
 
-		// alihkan halaman kembali
-		return redirect("admin/talent/list");
+        // alihkan halaman kembali
+        return redirect("admin/talent/list");
     }
-    
-    public function download(){
 
-        $file=public_path(). "/file_importExcel/template.xlsx";
+    public function download()
+    {
+
+        $file = public_path() . "/file_importExcel/template.xlsx";
 
         return Response::download($file);
     }
-
 }
